LayoutTests/svg/dynamic-updates/SVGFESpecularLightingElement-inherit-lighting-color-css-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the inherited 'lighting-color' attribute of the SVGFESpecularLightingElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL document.defaultView.getComputedStyle(specularElement).getPropertyValue('lighting-color') should be rgb(255, 0, 0). Was rgb(255, 255, 255).
FAIL document.defaultView.getComputedStyle(specularElement).getPropertyValue('lighting-color') should be rgb(173, 255, 47). Was rgb(255, 255, 255).
PASS successfullyParsed is true

TEST COMPLETE
